Personal Exceptions

+++ src/main/java/com/roninds/dscommerce/services/ProductService.java
@@ -3,17 +3,19 @@ package com.roninds.dscommerce.services;
 import com.roninds.dscommerce.dto.ProductDTO;
 import com.roninds.dscommerce.entities.Product;
 import com.roninds.dscommerce.repositories.ProductRepository;
+import com.roninds.dscommerce.services.exceptions.DatabaseException;
+import com.roninds.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.Optional;
-import java.util.List;
 
 @Service
 public class ProductService {
@@ -24,7 +26,7 @@ public class ProductService {
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
         Optional<Product> result = repository.findById(id);
-        Product product = result.get();
+        Product product = result.orElseThrow(() -> new ResourceNotFoundException("Recurso não enconttrado!"));
         return new ProductDTO(product);
     }
 
@@ -48,13 +50,17 @@ public class ProductService {
 
     @Transactional
     public  ProductDTO update(ProductDTO productDTO, Long id){
-        Product product = repository.getReferenceById(id);
+        try{
+            Product product = repository.getReferenceById(id);
 
-        copyDtoToEntity(productDTO, product);
+            copyDtoToEntity(productDTO, product);
 
-        Product productSaved = repository.save(product);
+            Product productSaved = repository.save(product);
 
-        return new ProductDTO(productSaved);
+            return new ProductDTO(productSaved);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
     }
 
     @Transactional
@@ -65,7 +71,16 @@ public class ProductService {
         product.setImgUrl(productDTO.getImgUrl());
     }
 
-    public void delete(Long id){
-        repository.deleteById(id);
+    @Transactional(propagation = Propagation.SUPPORTS)
+    public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 }
